feat(RUN-820): extract worker_support from worker.py
Move 7 symbols (parse_context_envelope, reconstruct_soul,
build_budgeted_history, build_scoped_state, _BLOCK_TYPE_MAP,
_resolve_block_soul, _create_block) to new worker_support.py.
No re-exports, no __getattr__ shim. Update test imports.
